演示如何正确规范地评测中文分词的准确率
