+++ src/test/java/com/hankcs/hanlp/model/perceptron/MSR.java
@@ -18,16 +18,17 @@ import com.hankcs.hanlp.utility.TestUtility;
  */
 public class MSR
 {
-    public static String TRAIN_PATH = "data/test/icwb2-data/training/msr_training.utf8";
-    public static String TEST_PATH = "data/test/icwb2-data/testing/msr_test.utf8";
-    public static String GOLD_PATH = "data/test/icwb2-data/gold/msr_test_gold.utf8";
-    public static String MODEL_PATH = "data/test/msr_cws";
-    public static String OUTPUT_PATH = "data/test/msr_output.txt";
-    public static String TRAIN_WORDS = "data/test/icwb2-data/gold/msr_training_words.utf8";
+    public static final String TRAIN_PATH = "data/test/icwb2-data/training/msr_training.utf8";
+    public static final String TEST_PATH = "data/test/icwb2-data/testing/msr_test.utf8";
+    public static final String GOLD_PATH = "data/test/icwb2-data/gold/msr_test_gold.utf8";
+    public static final String MODEL_PATH = "data/test/msr_cws";
+    public static final String OUTPUT_PATH = "data/test/msr_output.txt";
+    public static final String TRAIN_WORDS = "data/test/icwb2-data/gold/msr_training_words.utf8";
+    public static String SIGHAN05_ROOT;
 
     static
     {
-        String sighan05 = TestUtility.ensureTestData("icwb2-data", "http://sighan.cs.uchicago.edu/bakeoff2005/data/icwb2-data.zip");
+        SIGHAN05_ROOT = TestUtility.ensureTestData("icwb2-data", "http://sighan.cs.uchicago.edu/bakeoff2005/data/icwb2-data.zip");
         if (!IOUtil.isFileExisted(TRAIN_PATH))
         {
             System.err.println("请下载 http://sighan.cs.uchicago.edu/bakeoff2005/data/icwb2-data.zip 并解压为 data/test/icwb2-data");
